+++ data/normalized/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Latest Microsoft Security News Changes
 
-Generated: 2026-03-20T06:52:09.401Z
+Generated: 2026-03-21T06:39:17.349Z
 
 - 2026-07-01 • Microsoft Sentinel • Action required • Call to action: update queries and automation by July 1, 2026 - standardized account entity naming in incidents and alerts
 - 2026-07-01 • Microsoft Sentinel • GA • , Microsoft Sentinel will be supported in the Defender portal only, and any remaining customers using the Azure portal will be automatically redirected.
@@ -16,9 +16,9 @@ Generated: 2026-03-20T06:52:09.401Z
 - 2026-03-04 • Microsoft Defender for Cloud • Preview • On-demand malware scanning of Azure Files in Microsoft Defender for Storage
 - 2026-03-01 • Microsoft Defender for Cloud • Deprecation • with additional
 - 2026-03-01 • Microsoft Defender for Endpoint • GA • The proactive user containment (contain user) action as part of the predictive shielding feature is now generally available. This action infuses activity data with exposure data to identify exposed credentials at risk of being compromised and reused to conduct malicious activity.
-- 2026-03-01 • Microsoft Defender for Endpoint • Preview • New Microsoft Secure Score recommendations
+- 2026-03-01 • Microsoft Defender for Endpoint • Preview • Release - macOS
 - 2026-03-01 • Microsoft Defender for Endpoint • Preview • Microsoft Secure Score now includes new recommendations to help organizations reduce exposure to common attack techniques involving remote access and authentication relay attacks
 - 2026-03-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Release version 20.126012.15.0 released: See
+- 2026-03-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Release version 20.126012.17.0 released: See
 - 2026-03-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Block file transfer over RDP
 - 2026-03-01 • Microsoft Defender for Identity • Ikke oppgitt • Updates to Secure Score category calculations for increased accuracy
-- 2026-03-01 • Microsoft Defender XDR • GA • To improve accuracy and better protect organizational identities, we've made updates to the Secure Score category calculations. Some security recommendations categorized as